Consertando erros e testando a comunicacao entre servicos

+++ Servico-Produtos/src/main/java/br/com/servico/produtos/dtos/ProdutoResponseDTO.java
@@ -1,5 +1,7 @@
 package br.com.servico.produtos.dtos;
 
-public record ProdutoResponseDTO(Long id, String nome, String descricao, Double preco, Integer estoque, CategoriaResponseDTO categoria) {
+import java.math.BigDecimal;
+
+public record ProdutoResponseDTO(Long id, String nome, String descricao, BigDecimal preco, Integer estoque, CategoriaResponseDTO categoria) {
 
 }
